refactored todo and proejct controller

+++ controllers/projectsController.js
@@ -1,11 +1,13 @@
 const Project = require('../models/Project');
 const mongoose = require('mongoose');
 const expireTime = require('../config/anonymousDataExpireTime');
+
 /**
  * @description Get all projects from a user
  * @route GET /api/v1/projects
  * @access Private
  */
+
 const getAllProjects = async (req, res) => {
     const { uid } = req.user;
     const projects = await Project.find({ uid: uid });
@@ -17,6 +19,22 @@ const getAllProjects = async (req, res) => {
         return res.status(204).json({ msg: `No projects found with uid ${uid}` });
     }
     res.status(200).json(projects);
+};
+
+/**
+ * @description Get one project
+ * @route GET /api/v1/projects/:id
+ * @access Private
+ */
+
+const getProject = async (req, res) => {
+    const {
+        user: { uid },
+        params: { id: projectId },
+    } = req;
+
+    const project = await Project.findOne({ uid: uid, _id: projectId });
+    res.status(200).json(project);
 };
 
 /**
@@ -35,11 +53,12 @@ const createNewProject = async (req, res) => {
 
     const newProject = await Project.create({ uid, title, expireAt: isAnonymous ? expireTime() : null });
 
-    if (newProject) {
-        return res.status(201).json({ msg: 'New Project has been created' });
-    } else {
-        return res.status(400).json({ msg: 'Invalid project data' });
-    }
+    return res.status(201).json({ msg: 'New Project has been created' });
+    // if (newProject) {
+    // return res.status(201).json({ msg: 'New Project has been created' });
+    // } else {
+    //     return res.status(400).json({ msg: 'Invalid project data' });
+    // }
 };
 
 /**
@@ -49,20 +68,29 @@ const createNewProject = async (req, res) => {
  */
 
 const updateProject = async (req, res) => {
-    const { _id: projectId, title } = req.body;
+    const {
+        body: { title },
+        user: { uid },
+        params: { id: projectId },
+    } = req;
 
+    if (!title) {
+        return res.status(400).json({ msg: `No todo title` });
+    }
+    if (!uid) {
+        return res.status(400).json({ msg: `No uid` });
+    }
+    const project = await Project.findByIdAndUpdate({ _id: projectId, uid: uid }, req.body, {
+        new: true,
+        runValidators: true,
+    });
     // findById(): Finds a single document by its _id field
-    const project = await Project.findById(projectId).exec();
+    // const project = await Project.findById(projectId).exec();
     if (!project) {
         return res.status(400).json({ message: 'Project not found' });
     }
-    if (!title) {
-        return res.status(400).json({ msg: `No title` });
-    }
-    project.title = title;
-    const updatedProject = await project.save();
 
-    return res.status(200).json({ msg: `Project updated. id: ${updatedProject._id}` });
+    return res.status(200).json({ msg: `Project updated. id: ${project._id}` });
 };
 
 /**
@@ -72,24 +100,31 @@ const updateProject = async (req, res) => {
  */
 
 const deleteProject = async (req, res) => {
-    const { _id: projectId } = req.body;
-    // console.log('🚀 ~ file: projectsController.js:72 ~ deleteProject ~ req.body', req.body);
+    const {
+        user: { uid },
+        params: { id: projectId },
+    } = req;
 
     if (!projectId) {
         return res.status(400).json({ msg: 'ProjectId required' });
     }
+    if (!uid) {
+        return res.status(400).json({ msg: `No uid` });
+    }
 
-    const project = await Project.findById(projectId).exec();
+    const project = await Project.findByIdAndDelete({ _id: projectId, uid: uid });
+    // const project = await Project.findById(projectId).exec();
     if (!project) {
         return res.status(400).json({ msg: 'Project not found' });
     }
     const result = await project.deleteOne();
-    res.json({ msg: `Project with ID ${result._id} has been deleted` });
+    res.json({ msg: `Project has been deleted` });
 };
 
 module.exports = {
     createNewProject,
     getAllProjects,
+    getProject,
     updateProject,
     deleteProject,
 };
